% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{blast}
\alias{blast}
\title{Execute a given blast command}
\usage{
blast(command, output)
}
\arguments{
\item{command}{The blast command to be run}

\item{output}{The output file to store the blast results in}
}
\value{
The output file generated by the blastn call
}
\description{
Execute a given blast command
}
