% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emm.R
\name{emm}
\alias{emm}
\title{emm typing pipeline from WGS assemblies}
\usage{
emm(org_id, samples.df, locus)
}
\arguments{
\item{org_id}{Organism to query: GAS}

\item{samples.df}{Data frame of user selected samples and the sample paths}

\item{locus}{Sample number associated with contig.fasta file}
}
\value{
A table frame containing the results of the query
}
\description{
Takes Organism, Sample Number, Locus, and a Variable at queries a contig.fasta file
}
