% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_mlst.R
\name{general_mlst_pipeline}
\alias{general_mlst_pipeline}
\title{general_mlst_pipeline}
\usage{
general_mlst_pipeline(org_id, sample_num, locus, seq_type)
}
\arguments{
\item{org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{sample_num}{Sample number associated with contig.fasta file}

\item{locus}{The locus to query, or enter "list" to use a list of alleles}

\item{seq_type}{The sequence type to use. Can be MLST, NGMAST, or NSTAR}
}
\value{
A profile containing the sample number, and the loci tested on
}
\description{
Create a profile for MLST, NGMAST, or NGSTAR using blastn
}
\details{

}
