% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labware_dataframe.R
\name{get_labware_dataframe}
\alias{get_labware_dataframe}
\title{get_labware_dataframe.R}
\usage{
get_labware_dataframe()
}
\value{
A data frame
}
\description{
Creates and returns an empty data frame with headers following Labware guidelines
}
