% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samples.R
\name{get_samples}
\alias{get_samples}
\title{determine_sample}
\usage{
get_samples(org_id, sample_num)
}
\arguments{
\item{sample_num}{the sample the user input}
}
\value{
A data frame containing a read .csv file or containing a single sample
}
\description{
This function determines whether the user selected "list" or a specific sample number
and creates SampleList.df accordingly.
Used in: 23SrRNA.R, amr_databases.R, emm.R, master.R, pneumo_cat.R, vfdb.R
}
