% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Makeblastdb.R
\name{Index_pipeline}
\alias{Index_pipeline}
\title{Index BLAST fasta datases of lookup data}
\usage{
Index_pipeline(org_id, test_id, locus_id)
}
\arguments{
\item{org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{locus_id}{Locus_name to build blast database}

\item{TestID}{Sample number associated with contig.fasta file}
}
\value{
A table frame containing the results of the query
}
\description{
Takes Organism, Test, and Locus and create BLAST databases
}
