% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabwareUpload_GAS_TOXINS.R
\name{labware_gas_toxins}
\alias{labware_gas_toxins}
\title{run MasterBlastR first, then this script outputs csv to LabWareUpload_GAS_TOXINS.csv in Output folder}
\usage{
labware_gas_toxins()
}
\value{
A table frame containing the results of the query
}
\description{
Then run this analysis to combine data the full amr profile to upload to LabWare.
}
