% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabwareUpload_GONO_AMR.R
\name{labware_gono_amr}
\alias{labware_gono_amr}
\title{Run AMR first, then run the 23S allele counts, and then the NGSTAR-MLST analyses
Then run this analysis to combine data from AMR, 23S rRNA and NG-STAR
to prepare full amr profile to upload to LabWare.}
\usage{
labware_gono_amr()
}
\value{
A table frame containing the results of the query
}
\description{
Run AMR first, then run the 23S allele counts, and then the NGSTAR-MLST analyses
Then run this analysis to combine data from AMR, 23S rRNA and NG-STAR
to prepare full amr profile to upload to LabWare.
}
