% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf.R
\name{read_vcf}
\alias{read_vcf}
\title{Reads a .vcf file ignoring any headers, creates a data frame from the remaining lines}
\usage{
read_vcf(path)
}
\arguments{
\item{path}{The path to file corresponding .vcf file}
}
\value{
A Data Frame of the data from a .vcf
}
\description{
1. Grabs the header from the correct line
2. Creates a data frame from all found mutations
3. Change the column names of the data frame to the headers in 1.
}
