% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRNA.R
\name{rna_23s}
\alias{rna_23s}
\title{rRNA.R}
\usage{
rna_23s(org_id, samples)
}
\arguments{
\item{org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{samples}{Data frame of user selected samples and the sample paths}
}
\value{
A data frame containing the results of the query
}
\description{
rRNA.R
}
\details{
23S rRNA pipeline for WGS assemblies to determine number of mutated alleles

E.coli: A2059G and C2611T

GONO:
Run SNP core pipeline with NCCP11945_23S4.fasta file: A2045G or C2597T
PNEUMO:
Run SNP core pipeline with 23S_R6.fasta file: A2061G or C2613T

Takes Organism, Sample Number and queries a contig.fasta
Parses 23s rRNA mutations from VCF files

ON GALAXY:

Alternative_allele_proporition = 0.1
min_coverage = 15
min_mean_mapping = 30
run_name = 23S
in FreeBayes step of phylogeny set ploidy = 4

Export VCF from Galaxy:
Under "Collection Tools" , use the tool "Export to Warehouse" on "Filter vcf on collection 210".
It will bundle them together and makes a history item.  They are in /Warehouse/Temporary/galaxy_exports/RANDOM_NAME.
Click the "eye" to view RANDOM_NAME. You can then copy them from there using the regular file explorer into:
            " folder
}
