% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{ui}
\alias{ui}
\title{Function building dashboard UI, used in Shiny app}
\format{An object of class \code{shiny.tag} of length 3.}
\usage{
ui
}
\description{
Function building dashboard UI, used in Shiny app
}
\keyword{datasets}
