% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_pipeline.R
\name{database_pipeline}
\alias{database_pipeline}
\title{Generic Profiler : SampleNo_contigs.fasta vs. arg-annot/resfinder2/CARD or VFDB}
\usage{
database_pipeline(org_id, samples.df, is_vfdb, stdout = FALSE)
}
\arguments{
\item{org_id}{Organism to query: GAS or GONO}

\item{samples.df}{Data frame of user selected samples and the sample paths}

\item{is_vfdb}{Boolean, determine which database is being used. Specified from main call}

\item{stdout}{Boolean, write df to stdout? Specified from main call}
}
\value{
A table frame containing the results of the query
}
\description{
Generic Profiler : SampleNo_contigs.fasta vs. arg-annot/resfinder2/CARD or VFDB
}
\details{
GAS AMR loci
dfrF, dfrG, DHFR, ermA, ermB, ermT, folA, folP, gyrA, mefAE, msrD, parC, rpsJ, tetM, tetO
note that ermA == ermTR, ermC == ermT, mefA/E includes mefA and mefE; DHFR is the same as folA

ARG-ANNOT is a downloaded as a single multifasta

RESFINDER is separated into antimicrobial classes

VFDB is a downloaded from the web site as a single multifasta
}
