% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_mlst.R
\name{general_mlst_pipeline}
\alias{general_mlst_pipeline}
\title{Create a profile for MLST, NGMAST, or NGSTAR using blastn}
\usage{
general_mlst_pipeline(org_id, samples.df, locus, seq_type)
}
\arguments{
\item{org_id}{Organism to query: GAS, PNEUMO or GONO}

\item{samples.df}{Data frame of user selected samples and the sample paths}

\item{locus}{The locus to query, or enter "list" to use a list of alleles}

\item{seq_type}{The sequence type to use. Can be MLST, NGMAST, or NSTAR}
}
\value{
A profile containing the sample number, and the loci tested on
}
\description{
Create a profile for MLST, NGMAST, or NGSTAR using blastn
}
\details{

}
